refactor(backend): restructure authentication controller and move OAuth logic to separate controller
