--- apps/backend/src/modules/auth/dto/discord-signin.dto.ts
@@ -1,15 +1,0 @@
-import { IsString, IsNotEmpty } from 'class-validator';
-import { ApiProperty } from '@nestjs/swagger';
-
-/**
- * DTO for Discord Sign-In request
- */
-export class DiscordSignInDto {
-  @ApiProperty({
-    description: 'Discord access token received from Discord OAuth',
-    example: 'YOUR_DISCORD_ACCESS_TOKEN_HERE'
-  })
-  @IsString()
-  @IsNotEmpty()
-  accessToken!: string;
-}
