+++ apps/backend/src/modules/auth/dto/index.ts
@@ -1,7 +1,5 @@
 export * from './register-user.dto';
 export * from './login-user.dto';
 export * from './auth-response.dto';
-export * from './google-signin.dto';
-export * from './discord-signin.dto';
 export * from './link-google-account.dto';
 export * from './link-discord-account.dto';
