+++ apps/backend/src/modules/auth/index.ts
@@ -1,5 +1,7 @@
 export * from './auth.module';
 export * from './auth.controller';
+export * from './oauth.controller';
+export * from './account-linking.controller';
 export * from './dto';
 export * from './guards/jwt-auth.guard';
 export * from './strategies/jwt.strategy';
